+++ app/src/Middleware/PageConfig.php
@@ -19,6 +19,7 @@ class PageConfig extends Middleware
             }
             
             $this->view->getEnvironment()->addGlobal('pageConfig', $cfg);
+            $request->container->pageConfig = $cfg;
             return $next($request, $response);
         }